requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
dlib==19.24.2
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.67.1
h5py==3.12.1
idna==3.10
iso8601==2.1.0
Jinja2==3.1.4
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.9.0.80
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==5.28.3
Pygments==2.18.0
pyparsing==3.2.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
retina-face==0.0.14
rich==13.9.4
seaborn==0.13.2
serial==0.0.97
setuptools==75.3.0
six==1.16.0
soupsieve==2.6
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.1
Werkzeug==3.1.2
wheel==0.44.0
wrapt==1.16.0
